v3 value.hh: make isThunkOrApp underflow-trick unsigned (GCC -Werror=sign-compare)
The v3 OP_FORCE hot-path helper isThunkOrApp used (uint8_t - tApp) <= 1u, whose
result is a signed int compared to 1u — GCC/x86_64-linux -Werror=sign-compare
rejects it (clang/macOS is lenient).  Do the subtraction in unsigned so the
underflow intent (getInternalType() < tApp wraps to a large value > 1) is
explicit and warning-free.  Semantically identical.  First step toward an
x86_64-linux v3 build (the M5/cardano measurement host); the remaining blockers
are libstdc++-vs-libc++ missing-include debt + the flake not packaging
nix-expr-v3 (documented in memory).

Copyright (c) 2026 Moritz Angermann <moritz.angermann@iohk.io>, Input Output Group.
SPDX-License-Identifier: Apache-2.0

+++ src/libexpr/include/nix/expr/value.hh
@@ -1228,7 +1228,12 @@ public:
     /// compiler doesn't always CSE the two getInternalType() loads.
     inline bool isThunkOrApp() const noexcept
     {
-        return (static_cast<uint8_t>(getInternalType()) - tApp) <= 1u;
+        // Unsigned subtraction so the underflow trick (getInternalType() < tApp
+        // wraps to a large value > 1) is explicit — and GCC's -Werror=sign-compare
+        // doesn't fire on the `int <= 1u` it would otherwise see (clang/macOS is
+        // lenient; GCC/x86_64-linux is not).  Semantically identical.
+        return (static_cast<unsigned>(static_cast<uint8_t>(getInternalType()))
+                - static_cast<unsigned>(tApp)) <= 1u;
     }
 
     inline bool isBlackhole() const;
